#1363 Refactor transaction api package.
