+++ sharding-transaction/sharding-transaction-api/src/main/java/io/shardingsphere/transaction/manager/xa/XATransactionManager.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.manager;
+package io.shardingsphere.transaction.manager.xa;
 
 import io.shardingsphere.core.event.transaction.xa.XATransactionEvent;
 import io.shardingsphere.core.rule.DataSourceParameter;
