+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/XATransactionManagerSPILoader.java
@@ -17,6 +17,7 @@
 
 package io.shardingsphere.transaction.xa.manager;
 
+import io.shardingsphere.transaction.manager.xa.XATransactionManager;
 import io.shardingsphere.transaction.xa.manager.atomikos.AtomikosTransactionManager;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
