+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/atomikos/AtomikosTransactionManager.java
@@ -24,7 +24,7 @@ import io.shardingsphere.core.event.transaction.xa.XATransactionEvent;
 import io.shardingsphere.core.exception.ShardingException;
 import io.shardingsphere.core.rule.DataSourceParameter;
 import io.shardingsphere.transaction.xa.manager.XADataSourceWrapper;
-import io.shardingsphere.transaction.xa.manager.XATransactionManager;
+import io.shardingsphere.transaction.manager.xa.XATransactionManager;
 
 import javax.sql.DataSource;
 import javax.sql.XADataSource;
